Refactor QueueUtilsService and DownloadManager, added service and tests
