+++ src/app/network/DownloadManager.test.ts
@@ -1,0 +1,246 @@
+import { DownloadItem, DownloadManagerService, DownloadTask } from 'app/drive/services/downloadManager.service';
+import { createFilesIterator, createFoldersIterator } from 'app/drive/services/folder.service';
+import { DriveFileData, DriveFolderData, DriveItemData } from 'app/drive/types';
+import tasksService from 'app/tasks/services/tasks.service';
+import { QueueUtilsService } from 'app/utils/queueUtils';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { DownloadManager } from './DownloadManager';
+import { EncryptionVersion, FileStatus } from '@internxt/sdk/dist/drive/storage/types';
+
+describe('downloadManager', () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  it('should generate task for a folder and download it using the queue', async () => {
+    const mockFolder: DriveFolderData = {
+      id: 0,
+      uuid: 'uuid',
+      name: 'Folder1',
+      bucket: 'bucket',
+      parentId: 0,
+      parent_id: 0,
+      parentUuid: 'parentUuid',
+      userId: 0,
+      user_id: 0,
+      icon: null,
+      iconId: null,
+      icon_id: null,
+      isFolder: true,
+      color: null,
+      encrypt_version: null,
+      plain_name: 'Folder1',
+      deleted: false,
+      createdAt: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+    };
+    const downloadItem: DownloadItem = {
+      payload: [mockFolder as DriveItemData],
+      selectedWorkspace: null,
+      workspaceCredentials: null,
+    };
+    const mockTask: DownloadTask = {
+      items: downloadItem.payload,
+      credentials: {
+        credentials: {
+          user: 'any-user',
+          pass: 'any-pass',
+        },
+        mnemonic: 'any-mnemonic',
+      },
+      taskId: 'any-taskId',
+      createFilesIterator: createFilesIterator,
+      createFoldersIterator: createFoldersIterator,
+      options: {
+        areSharedItems: false,
+        downloadName: mockFolder.name,
+        showErrors: true,
+      },
+    };
+
+    vi.spyOn(DownloadManagerService.instance, 'generateTasksForItem').mockResolvedValue(mockTask);
+    vi.spyOn(QueueUtilsService.instance, 'getConcurrencyUsingPerfomance').mockReturnValue(6);
+    vi.spyOn(tasksService, 'addListener').mockReturnValue();
+    vi.spyOn(tasksService, 'removeListener').mockReturnValue();
+    vi.spyOn(tasksService, 'updateTask').mockReturnValue();
+    vi.spyOn(tasksService, 'findTask').mockReturnValue(undefined);
+
+    const downloadFolderSpy = vi.spyOn(DownloadManagerService.instance, 'downloadFolder').mockResolvedValue();
+    const downloadFileSpy = vi
+      .spyOn(DownloadManagerService.instance, 'downloadFile')
+      .mockRejectedValue(new Error('It should download folder'));
+    const downloadItemsSpy = vi
+      .spyOn(DownloadManagerService.instance, 'downloadItems')
+      .mockRejectedValue(new Error('It should download folder'));
+
+    await DownloadManager.add(downloadItem);
+
+    expect(downloadFolderSpy).toHaveBeenCalledOnce();
+    expect(downloadFolderSpy).toHaveBeenCalledWith(mockTask, expect.anything(), expect.anything());
+    expect(downloadFileSpy).not.toHaveBeenCalled();
+    expect(downloadItemsSpy).not.toHaveBeenCalled();
+  });
+
+  it('should generate task for a file and download it using the queue', async () => {
+    const mockFile: DriveFileData = {
+      id: 0,
+      uuid: 'uuid',
+      bucket: 'bucket',
+      name: 'File1',
+      plainName: 'File1',
+      plain_name: 'File1',
+      type: 'jpg',
+      size: 100,
+      fileId: 'fileId1',
+      folder_id: 1,
+      folderId: 1,
+      folderUuid: 'uuid1',
+      createdAt: new Date().toISOString(),
+      created_at: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+      deleted: false,
+      deletedAt: null,
+      currentThumbnail: null,
+      encrypt_version: EncryptionVersion.Aes03,
+      status: FileStatus.EXISTS,
+      thumbnails: [],
+    };
+    const downloadItem: DownloadItem = {
+      payload: [mockFile as DriveItemData],
+      selectedWorkspace: null,
+      workspaceCredentials: null,
+    };
+    const mockTask: DownloadTask = {
+      items: downloadItem.payload,
+      credentials: {
+        credentials: {
+          user: 'any-user',
+          pass: 'any-pass',
+        },
+        mnemonic: 'any-mnemonic',
+      },
+      taskId: 'any-taskId',
+      createFilesIterator: createFilesIterator,
+      createFoldersIterator: createFoldersIterator,
+      options: {
+        areSharedItems: false,
+        downloadName: mockFile.name,
+        showErrors: true,
+      },
+    };
+
+    vi.spyOn(DownloadManagerService.instance, 'generateTasksForItem').mockResolvedValue(mockTask);
+    vi.spyOn(QueueUtilsService.instance, 'getConcurrencyUsingPerfomance').mockReturnValue(6);
+    vi.spyOn(tasksService, 'addListener').mockReturnValue();
+    vi.spyOn(tasksService, 'removeListener').mockReturnValue();
+    vi.spyOn(tasksService, 'updateTask').mockReturnValue();
+    vi.spyOn(tasksService, 'findTask').mockReturnValue(undefined);
+
+    const downloadFolderSpy = vi
+      .spyOn(DownloadManagerService.instance, 'downloadFolder')
+      .mockRejectedValue(new Error('It should download file'));
+    const downloadFileSpy = vi.spyOn(DownloadManagerService.instance, 'downloadFile').mockResolvedValue();
+    const downloadItemsSpy = vi
+      .spyOn(DownloadManagerService.instance, 'downloadItems')
+      .mockRejectedValue(new Error('It should download file'));
+
+    await DownloadManager.add(downloadItem);
+
+    expect(downloadFileSpy).toHaveBeenCalledOnce();
+    expect(downloadFileSpy).toHaveBeenCalledWith(mockTask, expect.anything());
+    expect(downloadFolderSpy).not.toHaveBeenCalled();
+    expect(downloadItemsSpy).not.toHaveBeenCalled();
+  });
+
+  it('should generate task for items and download them together using the queue', async () => {
+    const mockFile: DriveFileData = {
+      id: 0,
+      uuid: 'uuid',
+      bucket: 'bucket',
+      name: 'File1',
+      plainName: 'File1',
+      plain_name: 'File1',
+      type: 'jpg',
+      size: 100,
+      fileId: 'fileId1',
+      folder_id: 1,
+      folderId: 1,
+      folderUuid: 'uuid1',
+      createdAt: new Date().toISOString(),
+      created_at: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+      deleted: false,
+      deletedAt: null,
+      currentThumbnail: null,
+      encrypt_version: EncryptionVersion.Aes03,
+      status: FileStatus.EXISTS,
+      thumbnails: [],
+    };
+    const mockFolder: DriveFolderData = {
+      id: 0,
+      uuid: 'uuid',
+      name: 'Folder1',
+      bucket: 'bucket',
+      parentId: 0,
+      parent_id: 0,
+      parentUuid: 'parentUuid',
+      userId: 0,
+      user_id: 0,
+      icon: null,
+      iconId: null,
+      icon_id: null,
+      isFolder: true,
+      color: null,
+      encrypt_version: null,
+      plain_name: 'Folder1',
+      deleted: false,
+      createdAt: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+    };
+    const downloadItem: DownloadItem = {
+      payload: [mockFile as DriveItemData, mockFolder as DriveItemData],
+      selectedWorkspace: null,
+      workspaceCredentials: null,
+    };
+    const mockTask: DownloadTask = {
+      items: downloadItem.payload,
+      credentials: {
+        credentials: {
+          user: 'any-user',
+          pass: 'any-pass',
+        },
+        mnemonic: 'any-mnemonic',
+      },
+      taskId: 'any-taskId',
+      createFilesIterator: createFilesIterator,
+      createFoldersIterator: createFoldersIterator,
+      options: {
+        areSharedItems: false,
+        downloadName: mockFile.name,
+        showErrors: true,
+      },
+    };
+
+    vi.spyOn(DownloadManagerService.instance, 'generateTasksForItem').mockResolvedValue(mockTask);
+    vi.spyOn(QueueUtilsService.instance, 'getConcurrencyUsingPerfomance').mockReturnValue(6);
+    vi.spyOn(tasksService, 'addListener').mockReturnValue();
+    vi.spyOn(tasksService, 'removeListener').mockReturnValue();
+    vi.spyOn(tasksService, 'updateTask').mockReturnValue();
+    vi.spyOn(tasksService, 'findTask').mockReturnValue(undefined);
+
+    const downloadFolderSpy = vi
+      .spyOn(DownloadManagerService.instance, 'downloadFolder')
+      .mockRejectedValue(new Error('It should download items'));
+    const downloadFileSpy = vi
+      .spyOn(DownloadManagerService.instance, 'downloadFile')
+      .mockRejectedValue(new Error('It should download items'));
+    const downloadItemsSpy = vi.spyOn(DownloadManagerService.instance, 'downloadItems').mockResolvedValue();
+
+    await DownloadManager.add(downloadItem);
+
+    expect(downloadItemsSpy).toHaveBeenCalledOnce();
+    expect(downloadItemsSpy).toHaveBeenCalledWith(mockTask, expect.anything(), expect.anything());
+    expect(downloadFolderSpy).not.toHaveBeenCalled();
+    expect(downloadFileSpy).not.toHaveBeenCalled();
+  });
+});
