+++ src/app/utils/queueUtils.ts
@@ -1,5 +1,7 @@
 export class QueueUtilsService {
-  static readonly getConcurrencyUsingPerfomance = (currentConcurrency: number, maxConcurrency: number): number => {
+  public static readonly instance: QueueUtilsService = new QueueUtilsService();
+
+  public readonly getConcurrencyUsingPerfomance = (currentConcurrency: number, maxConcurrency: number): number => {
     let newConcurrency = currentConcurrency;
     try {
       const memory = window?.performance?.memory;
